+++ packages/office-ui-fabric-react/src/components/Nav/Nav.styles.ts
@@ -1,0 +1,28 @@
+import { INavStyleProps, INavStyles } from './Nav.types';
+import {
+  IStyle,
+  ITheme,
+} from '../../Styling';
+
+export const getStyles = (
+  props: INavStyleProps
+): INavStyles => {
+  const {
+    className,
+    theme,
+  } = props;
+
+  const { palette, semanticColors } = theme;
+
+  return ({
+    root: [
+      'ms-Nav',
+      {
+        // Insert css properties
+
+      }
+    ],
+
+    // Insert className styles
+  });
+};
